refining user and album

+++ backend/app/Http/Controllers/UsersController.php
@@ -20,6 +20,12 @@ class UsersController extends Controller
         //
     }
 
+    /**
+     * get a list or a single user
+     *
+     * @param  string [$uuid] user uuid to return
+     * @return Response
+     */
     public function get (string $uuid= null) {
         $result= (is_null($uuid))?
             DB::table('users')->get():
@@ -30,6 +36,13 @@ class UsersController extends Controller
         return $result;
     }
 
+    /**
+     * create or update an user
+     *
+     * @param  Request $request request params
+     * @param  string  [$uuid]  user uuid to update
+     * @return Response
+     */
     public function save (Request $request, string $uuid= null)
     {
         $data= $request->all();
@@ -41,6 +54,9 @@ class UsersController extends Controller
             $data['id']= $uuid;
             $data['updated_at']= date('Y-m-d H:i:s');
             $action= 'update';
+            if (!Validation::isAdmin($request)) {
+                unset($data['role']);
+            }
         }
         if ($request->filled('password')) {
             $data['password']= md5($request->input('password'));
@@ -54,12 +70,29 @@ class UsersController extends Controller
                     (($result >= 1)? response()->json($data, 200): response('Nothing to update', 202));
     }
 
-    public function delete (string $uuid) {
+    /**
+     * delete an user
+     *
+     * @param  Request $request request params
+     * @param  string  $uuid    user uuid to delete
+     * @return Response
+     */
+    public function delete (Request $request, string $uuid) {
+        if (!Validation::isAdmin($request)) {
+            return response('User must be admin to delete users', 401);
+        }
+
         $result= DB::table('users')->where('id', $uuid)->delete();
 
         return ($result === 1)? response('User deleted', 200): response('User not found', 400);
     }
 
+    /**
+     * do a login
+     *
+     * @param  Request  $request username and password
+     * @return Response
+     */
     public function login (Request $request) {
         $result= DB::table('users')->where([
             'username'=> $request->input('username'),
@@ -72,21 +105,23 @@ class UsersController extends Controller
     }
 
     public function validation (Request $request) {
-        // $token= json_decode(base64_decode($request->input('token')));
-
-        // if ($token->time >= (time()+(12*60*60))) {
-        //     return response('', 403);
-        // }
-
-        // $result= DB::table('users')->where('id', $token->v4)->get();
-
         $result= Validation::token($request->input('token'));
 
-        $this->removePass($result);
+        if (is_null($result)) {
+            return response('No header sent, user not logged', 400);
+        }
 
-        return (count($result) === 0)? response('Token not found', 400): response()->json($result, 200);
+        $col= new Collection([$result]);
+
+        $this->removePass($col);
+        return response()->json($col, 200);
     }
 
+    /**
+     * remove password hash from result collection
+     *
+     * @param Collection &$result collection of users
+     */
     protected function removePass (Collection &$result) {
         foreach ($result as &$r) {
             unset($r->password);
